antes de quitar las tablas

+++ cuentame-core/src/main/java/es/alten/cuentame/dao/ActorRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.data.querydsl.binding.QuerydslBinderCustomizer;
 
-public interface ActorRepository extends ElvisBaseRepository<Actor, Long, QActor>,
+public interface ActorRepository extends ElvisBaseRepository<Actor, Integer, QActor>,
         JpaSpecificationExecutor<Actor>,
         QuerydslPredicateExecutor<Actor>,
         QuerydslBinderCustomizer<QActor> {
